move generateHexColor and title character to utils.js and update square-image accordingly
