+++ webofneeds/won-owner-webapp/src/main/webapp/app/components/square-image.js
@@ -2,43 +2,22 @@
 
 import angular from 'angular';
 
+import {
+    generateHexColor,
+    generateTitleCharacter,
+} from '../utils.js'
+
 function genComponentConf() {
     let template = `
                 <img class="image" ng-show="self.src" ng-src="{{self.src}}"/>
-                <div ng-attr-title="{{self.title}}" class="image" style="background-color: {{self.generateHexColor()}}" ng-show="!self.src"><!-- figure out some better way to color -->
-                    <span class="image__noimage">{{self.generateTitleCharacter()}}</span>
+                <div title="{{self.title}}" class="image" style="background-color: {{self.generateHexColor(self.uri)}}" ng-show="!self.src"><!-- figure out some better way to color -->
+                    <span class="image__noimage">{{self.generateTitleCharacter(self.title)}}</span>
                 </div>`;
 
     class Controller {
-        constructor() {}
-
-        generateHexColor(title) {
-            var hash = 0;
-
-            if(this.uri){
-                for (var i = 0; i < this.uri.length; i++) {
-                    hash = this.uri.charCodeAt(i) + ((hash << 5) - hash);
-                }
-            }
-
-            var c = (hash & 0x00FFFFFF)
-                .toString(16);
-
-            return "#"+("00000".substring(0, 6 - c.length) + c);
-        }
-
-        generateTitleCharacter() {
-            if(this.title) {
-                try {
-                    return this.title.charAt(0).toUpperCase();
-                } catch (err) {
-                    //for resilience purposes, since we can't be sure whether a need is a string or anything else
-                    console.warn("Title Character could not be retrieved from: ", this.title);
-                    return "?";
-                }
-            } else {
-                return "?";
-            }
+        constructor() {
+            this.generateHexColor = generateHexColor;
+            this.generateTitleCharacter = generateTitleCharacter;
         }
     }
